Major Refactoring! New configuration.py with the factories. Distinction between PlayerUI and PlayerLogic. Got rid off a dependency cycle. Movement of the player is in the Player's Logic, the PlayerUI just grabs the information where to draw of the logical Player class
