+++ .travis.yml
@@ -1,6 +1,7 @@
 language: python
 python:
   - "3.4"
+  - "2.7"
 sudo: required
 install:
   - sudo apt-get update -qq
