+++ setup.py
@@ -5,7 +5,7 @@ from setuptools import setup
 
 setup(name="tekmate",
       version="0.0.1",
-      description="Tekmate - Stargate-Based RPG",
+      description="Tekmate - Stargate-Based Point'n'Click",
       author="Max",
       packages=['tekmate'],
       license="GPLv3",
@@ -18,7 +18,8 @@ setup(name="tekmate",
           "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
           "Natural Language :: English",
           "Programming Language :: Python :: 3.4",
+          "Programming Language :: Python :: 2.7",
           "Topic :: Games/Entertainment",
           "Topic :: Software Development :: Libraries",
       ],
-)
+      )
