+++ tekmate/__main__.py
@@ -1,5 +1,6 @@
 # -*- encoding: utf-8 -*-
-from tekmate.game import PyGameInitializer, TekmateFactory
+
+from tekmate.configuration import PyGameInitializer, TekmateFactory
 
 
 def main():
@@ -7,6 +8,7 @@ def main():
     game_factory = TekmateFactory(initializer)
     game = game_factory.create()
     game.enter_mainloop()
+
 
 if __name__ == "__main__":
     main()
